+++ test_data.py
@@ -1,0 +1,2 @@
+EMAIL = 'test_andrew@test.ru'
+PASSWORD = 'Andrewkin'
