+++ tests/test_personal_account.py
@@ -1,14 +1,14 @@
-import time
-
 import allure
 from pages.login_page import LoginPage
 from pages.profile_page import ProfilePage
 from pages.main_page import MainPage
-from constants import ORDER_HISTORY_URL, PROFILE_PAGE
+from constants import ORDER_HISTORY_URL, PROFILE_PAGE, LOGIN_PAGE
 
 
 class TestPersonalAccount:
 
+
+    allure.title("Тестирование личного кабинета")
     def test_personal_account(self, browser):
         login_page = LoginPage(browser)
         profile_page = ProfilePage(browser)
@@ -19,8 +19,9 @@ class TestPersonalAccount:
             login_page.log_in()
         with allure.step("Нажимаем на Личный кабинет"):
             main_page.click_to_personal_account_link()
-        assert browser.current_url == PROFILE_PAGE
+        assert main_page.get_current_url() == PROFILE_PAGE
         with allure.step("Нажимаем на Историю заказов"):
             profile_page.click_history_order()
-        assert browser.current_url == ORDER_HISTORY_URL
+        assert main_page.get_current_url() == ORDER_HISTORY_URL
         profile_page.log_out()
+        assert main_page.get_current_url() == LOGIN_PAGE
